FIXUP: Silence most warnings
Silence MSVC bit shift warnings

A couple of files inside the partition allocator produce a ton of
the following MSVC warning: "warning C4334: '<<': result of 32-bit shift
implicitly converted to 64 bits (was 64-bit shift intended?)". This
commit explicitly marks the value 1 which is being bit-shifted as
a size_t, which silences those warnings.

Reviewed-on: https://codereview.qt-project.org/c/qt/qtwebengine-chromium/+/571231

+++ chromium/base/allocator/partition_allocator/src/partition_alloc/page_allocator_constants.h
@@ -123,7 +123,7 @@ PageAllocationGranularity() {
   }
   return size;
 #else
-  return 1 << PageAllocationGranularityShift();
+  return size_t(1) << PageAllocationGranularityShift();
 #endif
 }
